Убрать не нужные проверки на пустоту, вынести константы для описания функций, исправлена логика классов ChronotypeFunction и SleeplessNightsFunction переходы через ночь, ChronotypeName реализован а Chronotype, добавлен пограничный тест в ChronotypeFunctionTest на равенство OWL и LARK

+++ src/test/java/ru/yandex/practicum/sleeptracker/SleepTrackerAppTest.java
@@ -12,8 +12,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class SleepTrackerAppTest {
 
-    private static final long ZERO = 0;
-    private static final long ONE = 1;
+    private static final int ZERO = 0;
+    private static final int     ONE = 1;
 
     @Test
     void should_ReturnZeroForEmptyList() {
